move bertopic schemas into ppg-common
