+++ app/aimodels/bertopic/crud/crud_document_embedding_computation.py
@@ -1,6 +1,6 @@
+from ppg.schemas.bertopic.document_embedding_computation import DocumentEmbeddingComputationCreate, DocumentEmbeddingComputationUpdate
 from app.crud.base import CRUDBase
 from ..models.document_embedding_computation import DocumentEmbeddingComputationModel
-from ..schemas.document_embedding_computation import DocumentEmbeddingComputationCreate, DocumentEmbeddingComputationUpdate
 
 # CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType])
 class CRUDDocumentEmbeddingComputation(CRUDBase[DocumentEmbeddingComputationModel, DocumentEmbeddingComputationCreate, DocumentEmbeddingComputationUpdate]  ):
