--- app/aimodels/bertopic/schemas/__init__.py
@@ -1,4 +1,0 @@
-from .bertopic_embedding_pretrained import BertopicEmbeddingPretrained, BertopicEmbeddingPretrainedCreate, BertopicEmbeddingPretrainedInDB, BertopicEmbeddingPretrainedUpdate
-from .bertopic_trained import BertopicTrained, BertopicTrainedCreate, BertopicTrainedInDB, BertopicTrainedUpdate
-from .document import Document, DocumentCreate, DocumentInDB
-from .document_embedding_computation import DocumentEmbeddingComputation, DocumentEmbeddingComputationCreate, DocumentEmbeddingComputationInDB, DocumentEmbeddingComputationUpdate
