+++ ppg-common/ppg/schemas/bertopic/document.py
