+++ ppg-common/ppg/schemas/bertopic/document_embedding_computation.py
@@ -1,7 +1,8 @@
 from typing import Optional
+from pydantic import BaseModel, UUID4
+
 from ppg.core.config import OriginationEnum
 
-from pydantic import BaseModel, UUID4
 
 # Shared properties
 class DocumentEmbeddingComputationBase(BaseModel):
