+++ ppg-common/setup.py
@@ -7,6 +7,7 @@ setup(name='ppg-common',
       author='MIT Lincoln Laboratory',
       license='MIT',
       packages=['ppg.core',
+		'ppg.schemas.bertopic',
 		'ppg.schemas.mattermost',
 		'ppg.services'],
       include_package_data=True,
