+++ tests/aimodels/bertopic/routers/test_bertopic_embedding_pretrained.py
@@ -10,8 +10,8 @@ from sqlalchemy.orm import Session
 from unittest.mock import MagicMock
 from fastapi.testclient import TestClient
 from ppg.core.config import OriginationEnum
+from ppg.schemas.bertopic.bertopic_embedding_pretrained import BertopicEmbeddingPretrainedCreate
 from app.aimodels.bertopic.crud.crud_bertopic_embedding_pretrained import bertopic_embedding_pretrained
-from app.aimodels.bertopic.schemas.bertopic_embedding_pretrained import BertopicEmbeddingPretrainedCreate
 from app.aimodels.bertopic.ai_services.weak_learning import WeakLearner
 from fastapi.encoders import jsonable_encoder
