+++ tests/aimodels/bertopic/routers/test_documents.py
@@ -2,7 +2,7 @@ from fastapi.testclient import TestClient
 from fastapi.encoders import jsonable_encoder
 
 from ppg.core.config import OriginationEnum
-from app.aimodels.bertopic.schemas.document import DocumentCreate
+from ppg.schemas.bertopic.document import DocumentCreate
 
 
 def test_create_document_object_post_valid_request(client: TestClient):
